Move ITextJustifier into Text namespace for consistency
